perf: Reduce command line import time
