+++ dbgpt/model/cluster/worker/default_worker.py
@@ -19,12 +19,7 @@ from dbgpt.util.system_utils import get_system_info
 logger = logging.getLogger(__name__)
 
 _torch_imported = False
-try:
-    import torch
-
-    _torch_imported = True
-except ImportError:
-    pass
+torch = None
 
 
 class DefaultModelWorker(ModelWorker):
@@ -95,6 +90,8 @@ class DefaultModelWorker(ModelWorker):
     def start(
         self, model_params: ModelParameters = None, command_args: List[str] = None
     ) -> None:
+        # Lazy load torch
+        _try_import_torch()
         if not model_params:
             model_params = self.parse_parameters(command_args)
         self._model_params = model_params
@@ -436,3 +433,14 @@ def _new_metrics_from_model_output(
             ].available_memory_gb
 
     return metrics
+
+
+def _try_import_torch():
+    global torch
+    global _torch_imported
+    try:
+        import torch
+
+        _torch_imported = True
+    except ImportError:
+        pass
